requirements.txt
apt-get dependencies 

streamlit_lottie-dev
streamlit_option_menu-dev